basic CAN channel sampling working

+++ include/CAN/OBD2.h
@@ -30,7 +30,7 @@ CPP_GUARD_BEGIN
 #define OBD2_PID_DEFAULT_TIMEOUT_MS 300
 
 int OBD2_request_PID(unsigned char pid, int *value, size_t timeout);
-void OBD2_set_current_PID_value(size_t index, int value);
+void OBD2_set_current_PID_value(int index, int value);
 int OBD2_get_current_PID_value(int index);
 
 CPP_GUARD_END
